Fix import ordering issue
These two imports must strictly be imported in this order:

```html
<script type="module" src="./src/webcomponents/user-provider.ts"></script>
<script type="module" src="./src/webcomponents/site-nav.ts"></script>
```

Rollup (which vite uses) can reorder these in a non-deterministic way. This is an open issue and is not solved.  In our latest deployment it has broken.

This PR adds an import to site-nav.ts so that rollup will bundle these correctly.