services/annotations: check species before adding to annotation (#328)
* services/annotations: check species before adding to annotation

This was done to make a test pass. Namely TestAnnotationAddNonExistingSpeciesIdentification.
This can be skipped with the short flag but I didn't realise that.

* Update code-checks.yaml

This was done because I was getting this error: Error: Command pnpm recursive install (cwd: undefined) exits with status 1

* Update code-checks.yaml

Revert previous change, it didn't work.

* ui: replace GitHub path dependency with workspace protocol for @openfish/client

This was done because CI was failing, and this will hopefully fix it.

* services: use SpeciesExists helper function

This was done to simplify the code since I was previously doing it manually.