datastore: filestore field filtering beyond key-parts (#330)
* datastore: implement field filtering in the filestore

This was done so that we didn't have to design keys for the datastore based onthe limitations of our filestore.
Now the filestore will filter based on the keyparts if possible, but can parse the files to filter on other fields if necessary.

* datastore: add unit tests for FilterField

This was done to test the new field filtering functionality.

* datastore: add filestore comparisons for other datastore types

This includes string, int, float, time.

* cmd/openfish/services: return ErrNoSuchEntity if no user found by email

This was done because the TestGetUserByEmail function was panicing due to assuming this function returns a valid user or an error.